+++ src/EventHandlers/Projectors/Projector.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\EventSourcing\EventHandlers\Projectors;
+
+use Spatie\EventSourcing\EventHandlers\EventHandler;
+
+abstract class Projector implements EventHandler
+{
+    use ProjectsEvents;
+}