+++ tests/TestClasses/AutoDiscoverEventHandlers/TestQueuedProjector.php
@@ -2,10 +2,9 @@
 
 namespace Spatie\EventSourcing\Tests\TestClasses\AutoDiscoverEventHandlers;
 
-use Spatie\EventSourcing\Projectors\Projector;
-use Spatie\EventSourcing\Projectors\ProjectsEvents;
+use Spatie\EventSourcing\EventHandlers\Projectors\Projector;
+use Spatie\EventSourcing\EventHandlers\Projectors\ProjectsEvents;
 
-class TestQueuedProjector implements Projector
+class TestQueuedProjector extends Projector
 {
-    use ProjectsEvents;
 }