+++ tests/TestClasses/Projectors/ProjectThatHandlesASingleEvent.php
@@ -2,14 +2,12 @@
 
 namespace Spatie\EventSourcing\Tests\TestClasses\Projectors;
 
-use Spatie\EventSourcing\Projectors\Projector;
-use Spatie\EventSourcing\Projectors\ProjectsEvents;
+use Spatie\EventSourcing\EventHandlers\Projectors\Projector;
+use Spatie\EventSourcing\EventHandlers\Projectors\ProjectsEvents;
 use Spatie\EventSourcing\Tests\TestClasses\Events\MoneyAddedEvent;
 
-class ProjectThatHandlesASingleEvent implements Projector
+class ProjectThatHandlesASingleEvent extends Projector
 {
-    use ProjectsEvents;
-
     public string $handleEvent = MoneyAddedEvent::class;
 
     public function __invoke(MoneyAddedEvent $event)