+++ tests/TestClasses/Projectors/ResettableProjector.php
@@ -2,15 +2,13 @@
 
 namespace Spatie\EventSourcing\Tests\TestClasses\Projectors;
 
-use Spatie\EventSourcing\Projectors\Projector;
-use Spatie\EventSourcing\Projectors\ProjectsEvents;
+use Spatie\EventSourcing\EventHandlers\Projectors\Projector;
+use Spatie\EventSourcing\EventHandlers\Projectors\ProjectsEvents;
 use Spatie\EventSourcing\Tests\TestClasses\Events\MoneyAddedEvent;
 use Spatie\EventSourcing\Tests\TestClasses\Models\Account;
 
-class ResettableProjector implements Projector
+class ResettableProjector extends Projector
 {
-    use ProjectsEvents;
-
     protected array $handlesEvents = [
         MoneyAddedEvent::class => 'onMoneyAdded',
     ];