+++ tests/TestClasses/Reactors/BrokeReactor.php
@@ -5,13 +5,12 @@ namespace Spatie\EventSourcing\Tests\TestClasses\Reactors;
 use Illuminate\Support\Facades\Mail;
 use Spatie\EventSourcing\EventHandlers\EventHandler;
 use Spatie\EventSourcing\EventHandlers\HandlesEvents;
+use Spatie\EventSourcing\EventHandlers\Reactors\Reactor;
 use Spatie\EventSourcing\Tests\TestClasses\Events\MoneySubtractedEvent;
 use Spatie\EventSourcing\Tests\TestClasses\Mailables\AccountBroke;
 
-class BrokeReactor implements EventHandler
+class BrokeReactor extends Reactor
 {
-    use HandlesEvents;
-
     protected array $handlesEvents = [
         MoneySubtractedEvent::class => 'onMoneySubtracted',
     ];